Refactor form assets handling to separate class
[MAILPOET-2250]
